requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
lxml==4.5.0
# Editable Git install with no remote (mailgun-nextcloud==0.0.1)
-e /var/home/ramblurr/src/mailgun-nextcloud
MarkupSafe==1.1.1
mccabe==0.6.1
parso==0.6.1
pathspec==0.7.0
pycodestyle==2.5.0
pyflakes==2.1.1
python-dateutil==2.8.1
python-dotenv==0.11.0
regex==2020.1.8
requests==2.22.0
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.8
webdavclient3==3.14
Werkzeug==1.0.0
